+++ micropython/settings.py
@@ -48,7 +48,7 @@ FIELD_DATA_MEM_LOW_WATERMARK = 40 * 1024
 UNIT_ID = "None"
 UNIT_Name = "No Device Name"
 NODE_TYPE = 'base'
-FIRMWARE_VERSION = "v2.00.2w"
+FIRMWARE_VERSION = "v2.00.2x"
 
 # ============================================================
 # BOOTSTRAP CREDENTIALS (Required for First-Boot Provisioning)